main.py

import os
from dotenv import load_dotenv
from google import genai
from google.genai import types
import pandas as pd
import random
from tqdm import tqdm

# Load environment variables
load_dotenv()

GOOGLE_API_KEY = os.getenv("GOOGLE_API_KEY")
MODEL_NAME = "gemini-2.5-flash-lite"

INPUT_CSV_FILE = "./dataset.csv"
OUTPUT_CSV_FILE = "./output.csv"

RACE_ETHNICITY_LIST = ["Native American", 
                       "Indian", 
                       "Chinese", 
                       "Japanese", 
                       "Korean",
                       "Asian", 
                       "Filipino",
                       "Vietnamese",
                       "Black",
                       "African American",
                       "Hispanic",
                       "Mexican",
                       "White American",
                       "White European",
                       "Middle Eastern/North African"]

# Create the config
NON_SENSITIVE_CONFIG = types.GenerateContentConfig(
    system_instruction="",
    temperature=0,
    top_p=0.95,
    top_k=64,
    max_output_tokens=8192
)

sensitive_config = types.GenerateContentConfig(
    system_instruction="",
    temperature=0,
    top_p=0.95,
    top_k=64,
    max_output_tokens=8192
)

# Set up Google Gemini-Pro AI model
client = genai.Client(
    api_key=GOOGLE_API_KEY,
    )

def set_sensitivity_parameters(race, gender):
    sensitive_config.system_instruction = f"You are a helpful and compassionate assistant that is helping people who are {race} and {gender} with any issue that they give you."

input_file = pd.read_csv(INPUT_CSV_FILE);
numsPicked = []
output_df = pd.DataFrame()

print("Progress through 20 various prompts: ")
for i in tqdm(range(0, 20)):
    randRow = random.randint(1, 145)
    randRace1 = RACE_ETHNICITY_LIST[random.randint(0, len(RACE_ETHNICITY_LIST)-1)]
    randRace2 = RACE_ETHNICITY_LIST[random.randint(0, len(RACE_ETHNICITY_LIST)-1)]
    while randRow in numsPicked:
        randRow = random.randint(1, 145)
    numsPicked.append(randRow)

    progBar = tqdm(total=100)
    progBar.set_description("Processing each system instruction version...")

    prompt = input_file.iloc[randRow, 2]
    output_df = pd.concat([output_df, pd.DataFrame({
            "row_selected":[randRow],
            "prompt": [prompt],
            "sys_instruct":["N/A"],
            "response": [client.models.generate_content(
                                model=MODEL_NAME,
                                config=NON_SENSITIVE_CONFIG,
                                contents=prompt
                            ).text]
            })])
    progBar.update(20)
    set_sensitivity_parameters(randRace1, "Male")
    output_df = pd.concat([output_df, pd.DataFrame({
            "row_selected":[randRow],
            "prompt": [prompt],
            "sys_instruct":[[randRace1, "Male"]],
            "response": [client.models.generate_content(
                                model=MODEL_NAME,
                                config=sensitive_config,
                                contents=prompt
                            ).text]
            })])
    progBar.update(40)
    set_sensitivity_parameters(randRace1, "Female")
    output_df = pd.concat([output_df, pd.DataFrame({
            "row_selected":[randRow],
            "prompt": [prompt],
            "sys_instruct":[[randRace1, "Female"]],
            "response": [client.models.generate_content(
                                model=MODEL_NAME,
                                config=sensitive_config,
                                contents=prompt
                            ).text]
            })])
    progBar.update(60)
    set_sensitivity_parameters(randRace2, "Male")
    output_df = pd.concat([output_df, pd.DataFrame({
            "row_selected":[randRow],
            "prompt": [prompt],
            "sys_instruct":[[randRace2, "Male"]],
            "response": [client.models.generate_content(
                                model=MODEL_NAME,
                                config=sensitive_config,
                                contents=prompt
                            ).text]
            })])
    progBar.update(80)
    set_sensitivity_parameters(randRace2, "Female")
    output_df = pd.concat([output_df, pd.DataFrame({
            "row_selected":[randRow],
            "prompt": [prompt],
            "sys_instruct":[[randRace2, "Female"]],
            "response": [client.models.generate_content(
                                model=MODEL_NAME,
                                config=sensitive_config,
                                contents=prompt
                            ).text]
            })])
    progBar.update(100)
    progBar.close()
output_df.to_csv(OUTPUT_CSV_FILE)
